[BH-387] Move utils/state to service-antenna
utils/state is only one header used in one service,
moving it to service-antenna
