+++ module-utils/CMakeLists.txt
@@ -32,7 +32,6 @@ set (SOURCES
         ${CMAKE_CURRENT_SOURCE_DIR}/PhoneNumber.hpp
         ${CMAKE_CURRENT_SOURCE_DIR}/NumberHolderMatcher.hpp
         ${CMAKE_CURRENT_SOURCE_DIR}/country.hpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/state/ServiceState.hpp
         ${CMAKE_CURRENT_SOURCE_DIR}/ical/ParserICS.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/time/TimeRangeParser.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/time/DateAndTimeSettings.cpp
